fix: Isolate the creation of the PlatformOptions and InitializeOptions into a ConfigurationUtility class.
