+++ com.playeveryware.eos/Runtime/Core/IPlatformSpecifics.cs
@@ -35,7 +35,7 @@ namespace PlayEveryWare.EpicOnlineServices
 
         string GetDynamicLibraryExtension();
 
-//#if EOS_DYNAMIC_BINDINGS
+        //#if EOS_DYNAMIC_BINDINGS
         // Only called if EOS_DYNAMIC_BINDINGS is defined
         void LoadDelegatesWithEOSBindingAPI();
         //#endif
@@ -45,10 +45,10 @@ namespace PlayEveryWare.EpicOnlineServices
 #if !EXTERNAL_TO_UNITY
         void ConfigureSystemInitOptions(ref EOSInitializeOptions initializeOptions);
 
-        void ConfigureSystemPlatformCreateOptions(ref EOSCreateOptions createOptions);
-
         void InitializeOverlay(IEOSCoroutineOwner owner);
 #endif
+
+        void ConfigureSystemPlatformCreateOptions(ref EOSCreateOptions createOptions);
 
         void RegisterForPlatformNotifications();
