+++ com.playeveryware.eos/Runtime/Core/Utility/ConfigurationUtility.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: a2a80b4f410801c4a9bc43800b725258
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
